pyproject.toml
[tool.poetry]
name = "web3_test_for_demo"
version = "0.1.0"
description = ""
authors = ["Gwendolyn Faraday <gwenfaraday@pm.me>"]

[tool.poetry.dependencies]
python = "^3.9"
web3 = "^5.28.0"
py-solc-x = "^1.1.1"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
